// app/src/types/cursors.tsx
export enum CursorNameEnum {
  Default = "cursor-default",
  Pointer = "cursor-pointer",
  Crosshair = "cursor-crosshair",
  Move = "cursor-move",
  Grab = "cursor-grab",
  Grabbing = "cursor-grabbing",
  Text = "cursor-text",
  NotAllowed = "cursor-not-allowed",
  EResize = "cursor-e-resize",
  NResize = "cursor-n-resize",
  NeResize = "cursor-ne-resize",
  NwResize = "cursor-nw-resize",
  SResize = "cursor-s-resize",
  SeResize = "cursor-se-resize",
  SwResize = "cursor-sw-resize",
  WResize = "cursor-w-resize",
  NsResize = "cursor-ns-resize",
  NeswResize = "cursor-nesw-resize",
  NwseResize = "cursor-nwse-resize",
  ColResize = "cursor-col-resize",
  RowResize = "cursor-row-resize",
  AllScroll = "cursor-all-scroll",
  ZoomIn = "cursor-zoom-in",
  ZoomOut = "cursor-zoom-out",
  GrabHand = "cursor-grab-hand",
  NotAllowedHand = "cursor-not-allowed-hand",
  Pen = "cursor-pen",
  Eraser = "cursor-eraser",
  Handwriting = "cursor-handwriting",
  ZoomInHand = "cursor-zoom-in-hand",
  ZoomOutHand = "cursor-zoom-out-hand",
  HandPointing = "cursor-hand-pointing",
  HandPointingRight = "cursor-hand-pointing-right",
  HandPointingLeft = "cursor-hand-pointing-left",
  HandPointingUp = "cursor-hand-pointing-up",
  HandPointingDown = "cursor-hand-pointing-down",
  HandPointingMiddle = "cursor-hand-pointing-middle",
  HandPointingRightMiddle = "cursor-hand-pointing-right-middle",
  HandPointingLeftMiddle = "cursor-hand-pointing-left-middle",
  HandPointingUpMiddle = "cursor-hand-pointing-up-middle",
  HandPointingDownMiddle = "cursor-hand-pointing-down-middle",
  HandPointingDouble = "cursor-hand-pointing-double",
  HandPointingDoubleRight = "cursor-hand-pointing-double-right",
  HandPointingDoubleLeft = "cursor-hand-pointing-double-left",
  HandPointingDoubleUp = "cursor-hand-pointing-double-up",
  HandPointingDoubleDown = "cursor-hand-pointing-double-down",
  HandPointingTriple = "cursor-hand-pointing-triple",
  HandPointingTripleRight = "cursor-hand-pointing-triple-right",
  HandPointingTripleLeft = "cursor-hand-pointing-triple-left",
  HandPointingTripleUp = "cursor-hand-pointing-triple-up",
  HandPointingTripleDown = "cursor-hand-pointing-triple-down",
  HandPointingQuadruple = "cursor-hand-pointing-quadruple",
  HandPointingQuadrupleRight = "cursor-hand-pointing-quadruple-right",
  HandPointingQuadrupleLeft = "cursor-hand-pointing-quadruple-left",
  HandPointingQuadrupleUp = "cursor-hand-pointing-quadruple-up",
  HandPointingQuadrupleDown = "cursor-hand-pointing-quadruple-down",
  HandPointingQuintuple = "cursor-hand-pointing-quintuple",
  HandPointingQuintupleRight = "cursor-hand-pointing-quintuple-right",
  HandPointingQuintupleLeft = "cursor-hand-pointing-quintuple-left",
  HandPointingQuintupleUp = "cursor-hand-pointing-quintuple-up",
  HandPointingQuintupleDown = "cursor-hand-pointing-quintuple-down",
  HandPointingSextuple = "cursor-hand-pointing-sextuple",
  HandPointingSextupleRight = "cursor-hand-pointing-sextuple-right",
  HandPointingSextupleLeft = "cursor-hand-pointing-sextuple-left",
  HandPointingSextupleUp = "cursor-hand-pointing-sextuple-up",
  HandPointingSextupleDown = "cursor-hand-pointing-sextuple-down",
  HandPointingSeptuple = "cursor-hand-pointing-septuple",
  HandPointingSeptupleRight = "cursor-hand-pointing-septuple-right",
  HandPointingSeptupleLeft = "cursor-hand-pointing-septuple-left",
  HandPointingSeptupleUp = "cursor-hand-pointing-septuple-up",
  HandPointingSeptupleDown = "cursor-hand-pointing-septuple-down",
  HandPointingOctuple = "cursor-hand-pointing-octuple",
  HandPointingOctupleRight = "cursor-hand-pointing-octuple-right",
  HandPointingOctupleLeft = "cursor-hand-pointing-octuple-left",
  HandPointingOctupleUp = "cursor-hand-pointing-octuple-up",
  HandPointingOctupleDown = "cursor-hand-pointing-octuple-down",
  HandPointingNonuple = "cursor-hand-pointing-nonuple",
  HandPointingNonupleRight = "cursor-hand-pointing-nonuple-right",
  HandPointingNonupleLeft = "cursor-hand-pointing-nonuple-left",
  HandPointingNonupleUp = "cursor-hand-pointing-nonuple-up",
  HandPointingNonupleDown = "cursor-hand-pointing-nonuple-down",
  HandPointingDecuple = "cursor-hand-pointing-decuple",
  HandPointingDecupleRight = "cursor-hand-pointing-decuple-right",
  HandPointingDecupleLeft = "cursor-hand-pointing-decuple-left",
  HandPointingDecupleUp = "cursor-hand-pointing-decuple-up",
  HandPointingDecupleDown = "cursor-hand-pointing-decuple-down",
  HandPointingUndecuple = "cursor-hand-pointing-undecuple",
  HandPointingUndecupleRight = "cursor-hand-pointing-undecuple-right",
  HandPointingUndecupleLeft = "cursor-hand-pointing-undecuple-left",
  HandPointingUndecupleUp = "cursor-hand-pointing-undecuple-up",
  HandPointingUndecupleDown = "cursor-hand-pointing-undecuple-down",
  HandPointingDuodecuple = "cursor-hand-pointing-duodecuple",
}